feat: 버전 복원 Toast 알림 및 에러 배너 추가
- RestoreToast 컴포넌트 생성 (블랙/화이트 무채색 디자인)
- Confetti 효과와 Liquid Glass 스타일 적용
- i18n 적용으로 실시간 언어 전환 지원
- 버전 복원 실패 시 에러 배너 표시
- alert 대신 Toast로 UX 개선
